requirements.txt
streamlit
biopython
transformers
torch
python-dotenv
sacremoses
